// src/components/Interface.tsx
export interface CoinHistoricalInfo {
  id?: string;
  symbol?: string;
  name?: string;
  localization?: any;
  image?: any;
  market_data?: any;
  community_data?: any;
  developer_data?: any;
  public_interest_stats?: any;
}

export interface CoinInfo {
  id: string;
  symbol: string;
  name: string;
  platforms: any;
}